consolidate PlantDisplay
