+++ src/routes/app/gallery/+page.svelte
@@ -122,7 +122,7 @@
   >
     {data.garden.name}
   </div>
-  {#each dataWithTimes.garden.plantsInGarden as plant}
+  {#each dataWithTimes.garden.plantsInGarden as plant, index}
     <!-- svelte-ignore a11y-click-events-have-key-events -->
     <!-- svelte-ignore a11y-no-static-element-interactions -->
     <!-- <code>{plant.plant.timeLeft.toFormat("mm:ss")}</code> -->
@@ -134,7 +134,7 @@
         }}
         class="cursor-pointer mt-4"
       >
-        <PlantDisplay {plant} applyFilters={false} />
+        <PlantDisplay imageUrl={plant.imageUrl || ""} applyFilters={false} />
       </div>
       <div class="mt-4 text-center">
         <button
@@ -154,7 +154,7 @@
         }}
         class="cursor-pointer mt-4"
       >
-        <PlantDisplay {plant} applyFilters={true} />
+        <PlantDisplay imageUrl={plant.imageUrl || ""} applyFilters={true} />
       </div>
       <div class="mt-4 text-center">
         <button
